# specs/001-multi-tenant-base/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Komanda Core API
  version: 1.0.0
  description: >
    Versioned contracts for tenant provisioning, catalog, storefront checkout,
    orders, Mercado Pago connection, webhooks, and print agents. The Astro-based
    komanda-business client owns plan selection, registration and verification UX;
    Core validates its planId against migration-managed definitions, owns atomic
    provisioning, issues and consumes identity-verification challenges, and enforces
    entitlement snapshots. Mercado Pago onboarding is OAuth-only.
    A path tenant identifier is never sufficient authorization; Core verifies
    membership or the appropriate integration/agent identity.
servers:
  - url: https://app.komanda.com
...
paths:
...
  /api/v1/auth/onboarding-handoffs/consume:
    post:
      tags: [Auth]
      operationId: consumeOnboardingHandoff
      description: >
        Atomically consumes the short-lived handoff after Core email
        verification and establishes the owner's revocable Core session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingHandoffConsumeRequest"
      responses:
        "204":
          description: Handoff consumed and secure session cookie set.
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: komanda_session
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-user-session
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-agent-token
...
  responses:
...
    Conflict:
      description: State, idempotency or relationship conflict.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ValidationError:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, correlationId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        correlationId:
          type: string
          format: uuid
        errors:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
...
    OnboardingHandoffConsumeRequest:
      type: object
      additionalProperties: false
      required: [tenantId, token]
      properties:
        tenantId:
          type: string
          format: uuid
        token:
          type: string
          minLength: 32
...